requirements.txt
accelerate==1.1.1
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudscraper==1.2.71
curl_cffi==0.7.4
diffusers==0.31.0
einops==0.8.0
filelock==3.16.1
freeGPT==1.3.5
freeGPTFix==1.3.0
frozenlist==1.5.0
fsspec==2024.10.0
gpt4free==0.0.0
huggingface-hub==0.26.3
idna==3.10
importlib_metadata==8.5.0
ip-adapter==0.1.0
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pillow==11.0.0
propcache==0.2.1
psutil==6.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
SQLAlchemy==2.0.36
sympy==1.13.1
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.3
zipp==3.21.0
